Remove explicit configuration of LLVM_ENABLE_ASSERTIONS
Summary:
Per llvm/docs/CMake.rst, LLVM_ENABLE_ASSERTIONS defaults to on
when build type is Debug. As a result, this explicit configuration is
unnecessary. Checking the history of Hermes code, at one point, assertions
when enabled unconditionally. When this behavior is no longer wanted,
the explicit configuration is left over, probably as an artifact.

Reviewed By: tmikov

Differential Revision: D14594563

fbshipit-source-id: ec1cb3fc493331b9b2f13793e048840aead21f07