Increase threshold in InterpreterTest.FrameSizeTest in Windows
Summary:
The tests asserts that Interpreter loop does not have a stack that is too deep. MSVC compiler allocates a much larger stack than gcc/clang does. The threshold is adjusted accordingly. There is no short term plan to understand the cause and resolve this issue.

The approach of increasing the threshold is the outcome of discussion in the associated task.

Reviewed By: tmikov

Differential Revision: D14594566

fbshipit-source-id: 5fbbfe7251de42a07973dc4703fcbd12f73dc90c