Move <global> source mapping logic to IR generation step
Summary: Previously the virtual source file path `"<global>"` was missing at source map population time, causing a crash when trying to create a mapping associated with that source. It seems that `"<global>"` was being added in `createSourceMapGenerator` but clobbered by `generateIRForSourcesAsCJSModules`. This change moves this logic into the latter function.

Reviewed By: avp

Differential Revision: D14667081

fbshipit-source-id: ac21f84acebaa50d7037e4d0c75b39da60a9598c