Add VMExperimentFlags to Runtime
Summary:
Add a new field 'VMExperimentFlags' to RuntimeConfig in Hermes runtime.

The field is an integer and the name of the field is intentionally general so it can
be reused for any VM experiments that needs to pass a value to the runtime
to conduct an experiment. This avoids the boilerplate that people need to add
and clean up for every VM experiment they need to do.

Reviewed By: ridiculousfish

Differential Revision: D14587127

fbshipit-source-id: 7627e5aee87a51d0a50d95e83071c74ce3ab3c81