Factor out stats code in more generally-usable form.
Summary:
In reviewing D14676416, Ashok pointed out that I was adding a new version of code extremely similar to a class he had written in Hermes for accumulating GC stats.

This diff factors that code out, and generalizes it somewhat, in xplat Hermes.  If this is approved, I will copy the new file to fbcode hermes, and use it in an update to D14676416.

Differential Revision: D14697438

fbshipit-source-id: 58e4896e21d277c9f7303c0f86ccf8c9793bbfcd