clean up NewExpression, MemberExpression, CallExpression
Summary:
Clean up and simplify the handling of NewExpression, MemberExpression
and CallExpression as it relates to the ambiguous handling of `new` and
call arguments.

The key realizations are that NewExpression can be one or more
constructor calls without arguments (the s-expressions below represent
productions that have been recursively matched, not AST nodes):

    new foo ->
        (NewExpression (MemberExpression))
    new new foo ->
        (NewExpression (NewExpression (MemberExpression)))

MemberExpression can be one or more constructor calls with arguments:

    new foo() ->
        (MemberExpression (MemberExpression))
    new new foo()()
        (MemberExpression (MemberExpression (MemberExpression))

And call expression are arguments that don't match up with a `new`:

    foo() ->
        (CallExpression (MemberExpression))
    new foo()()
        (CallExpression (MemberExpression (MemberExpression)))

Add a new test.

Reviewed By: avp

Differential Revision: D14595341

fbshipit-source-id: 31cbc0eb18f226ac6c2e70a13ff0294397ee3f4b