initialize arguments in Function.prototype.apply()
Summary:
Function.prototype.apply() did not initialize the outgoing argument
registers before starting a loop to populate them. Under certain
conditions that loop might cause allocations or arbitrary code to be
run, causing a garbage collection, which would see uninitialized values
in the stack.

The fix is simple: initialize the arguments before the loop.

Add a test which crashed before the fix and works after it.

Reviewed By: amnn

Differential Revision: D14768043

fbshipit-source-id: f5b11508d5a851404f992575946ce5099dcdba65