CopyableVector to support capacities of 0
Summary:
Prior to this fix, CopyableVector would initialize itself with a capacity
of at least 1. This was reasonable because the only way to append to it was
with push_back(). However with the new assignment operator, we want to allow
empty CopyableVectors that are populated later.

Support a capacity of 0, in which case the storage is a null pointer.
Also factor out the growth strategy to reduce duplication.

Reviewed By: avp

Differential Revision: D14686838

fbshipit-source-id: c7c0ebbf0ad3e71a3b8544bf602bb426224343ae