Clean up overflow string table entries
Summary:
Prior to this fix, the string table was represented via:

1. A count of SmallStringTableEntries and a byte count of the string table
itself
2. Overflow entries were appended after the SmallStringTableEntries

This required some tricky pointer arithmetic to encode/decode and makes it
harder to split these segments.

With this fix, the overflow string table is promoted to a real segment. We
no longer have a "string table segment size", just counts for small and
overflow entries. This is simpler to encode and decode.

Reviewed By: kodafb

Differential Revision: D14727874

fbshipit-source-id: 2369783c859c4960ce7a8f3ac808ff05e95fb546