Fix bug in serialization of MarkerRecord
Summary:
A plain `MarkerRecord` wasn't being serialized correctly, because it assumed that it was
abstract and there would be no instance of `MarkerRecord`.
This grew out of the need to inherit from both `MarkerRecord` and `ReturnRecord`.

So I changed `ReturnRecord` to not inherit from `Record` and made it a Mixin-class instead.
Now all uses of `ReturnRecord` are basically composition rather than inheritance.

Differential Revision: D14762417

fbshipit-source-id: 527f8b26b997da8ae65e6a796c36608f71d8de4a