Redo optimization in BCProviderFromSrc::createBCProviderFromSrc() diff in xplat hermes.
Summary:
Use a build mode flag to determine whether the compileJS API can do optimization.

Facebook

Jeffrey originally made these changes in fbcode hermes, allowing optimized compiles.  He ported these to xplat hermes, so they wouldn't go away in the next sync.  But he had to back these out in D14788203, because it increased the Runtime's code size by bringing in the optimizer portion of the compiler.

This diff does these modifications again, but this time puts them under a flag which is not defined by default.  So xplat should not change at all.  We can set this flag in a TARGETS file in fbcode hermes, so we get the needed capability.

Differential Revision: D14788203

fbshipit-source-id: dab1018ff3bac8edff5f7102dff0dcd86f61e101