Hold references to persistent BCProviders in Runtime
Summary:
BCProviders may indirectly be keeping alive memory which backs the string data in the identifier table. This data can be used across invocations in the same Runtime, so BCProviders need to be held until Runtime is destroyed.

Facebook

The specific motivating case here was HaaS which uses an mmap'd file buffer via JSI, and evaluates multiple JS sources during the runtime's lifetime. This triggered a segfault on later access to the identifier table.

Reviewed By: tmikov

Differential Revision: D14866080

fbshipit-source-id: 741e8e5de6dd6a5d016070c80bcb43b947d1c579