Add further checking to OrderedHashMap.
Summary:
One more level of checking in OrderedHashMap.

So far, we know from Android minidumps that an OHM ends up with hashTable_ and capacity_ set to zero.  We don't know whether this is part of a general zeroing of a section of memory, failure to update a pointer during GC, so it points to some zeroed section of memory, or something we haven't seen in the implementation that allows this state to be reached...or something else.

This diff checks some parts of the well-formedness of OHMs on entry and exit of important methods.  In particular, it verifies that the vtable pointer, the first field of a GCCell, has a valid value.  The hashTable_ field that gets zero'd is second; capacity_ is 4th or 5th.  So if this were zero, we'd have stronger evidence for some general zeroing mechanism.  On the other hand, if the GCCell is still a valid OHM, then we'd have to look more strongly at mechanisms within the implementation of OHM.

Reviewed By: kodafb

Differential Revision: D14893486

fbshipit-source-id: 9e1a6901b6a3723a057444ad0eac81b9afdd3a93