Add mitigation for detached ArrayBuffer to TypedArray::getOwnIndexed
Summary:
If a TypedArray is viewing a detached ArrayBuffer, it should *not* attempt to read any values from memory.
Right now what will happen is that `src_` still points to a potentially de-allocated region of memory, and `getOwnIndexed`
will attempt to read from it, possibly returning garbage values.

The spec says that this should throw `TypeError`, but it's not feasible to do that now. Instead, return the number `0`.
I chose not to return `undefined` because it has the potential to break tests such as `toLocaleString`.

Reviewed By: avp

Differential Revision: D14902392

fbshipit-source-id: 5d2eb4ff30f240ceaa5ce2c03fcc4754d637e2d1