Fix bug in NCGen handlesan.
Summary: This is a bug in the NCGen implementation of HandleSan itself.  NCGen has the concept of the current AllocContext; this is "borrowed" either from the current alloc segment of the young or old generations.  Whenever we do a collection, NCGen has to "yield" the current AllocContext back to its owner, do the collection, then claim it back.  We weren't doing that when we did HandleSan collections, causing odd behavior.

Reviewed By: tmikov

Differential Revision: D14928847

fbshipit-source-id: 8de776e81aa092efff4c04ee88e99af94e7cf815