Disable ArrayStorage::compaction (= trimming) to fix bug.
Summary:
We finally have a diagnosis for the "OrderedHashMap bug."   It turns out to be caused (we sincerely hope) by "ArrayStorage trimming."  OrderedHashMap::create allocates an ArrayStorage of the required initial capacity to be the hashTable_ -- with initial size zero.  If then allocates memory for the OrderedHashMap.  If this second allocation causes a full collection, ArrayStorage trimming decreases the capacity of the ArrayStorage to its size (zero).  It then calls the ArrayStorage constructor on the memory allocated for the OrderedHashMap.  This, and other operations on the OrderedHashMap, assume that the ArrayStorage has size/capacity equal to the capacity_ of the OrderedHashMap -- which means it writes outside the bounds of the ArrayStorage object, causing havoc.

This diff #if's out the use of this feature, marking the code with a TODO referencing a task to decide the fate of this trimming feature.

Reviewed By: tmikov

Differential Revision: D14928863

fbshipit-source-id: 15b160424843670970d2643c069a5950b15801d2