Make ConsecutiveStringStorage::getStringTableView more permissive.
Summary:
Previously it returned an ArrayRef which did not permit mutation.  In a later
diff, I need some way to re-order the string table's index and I chose to do it
by obtaining a reference through this API.

This diff also removes `markEntryAsIdentifier` as this can now be done by
accessing the appropriate entry through the view and marking it.

Reviewed By: avp

Differential Revision: D14970871

fbshipit-source-id: 3d8c16007ecd45c0b3edae91ff783e4f5e2f63ff