Extract SetVector logic out of UniquingStringTable
Summary:
I've been trying to update `UniquingStringTable`'s API to accommodate the new
ways it will be used for capturing string literals.  Namely:

 - It should always gather all the strings ahead of time, rather than during
   ISel.  Currently it only does this when optimizations are enabled and there
   is no base bytecode bundle.
 - It should keep track of which strings are identifiers and which are not
   during this pre-ISel traversal.
 - Its API should not expose the IDs of strings during the traversal.
 - It should pack the string storage (in opt mode) before ISel.

All so that the string index can be re-ordered, and then remain unchanged
during ISel where we simply look up IDs to existing literals and identifiers.

One constant stumbling block is that `UniquingStringTable` is also used to
gather filenames for debug info and in this capacity it has a much simpler
usage:  They are gathered lazily during ISel and there is no reason to do this
all in one go ahead of time.

I want to separate `UniquingStringTable` into `UniquingStringLiteralTable` and
`UniquingFilenameTable` to prevent these two use cases from interfering with
each other.

This diff is the first step:  Extracting what is shared between them into a new
class -- `StringSetVector` -- which is responsible for accumulating unique
strings in first insertion order.

I also took this opportunity to remove

```
std::vector<llvm::StringRef> stringsByIndex_
```

It was not necessary as the owning storage held all the strings in the correct
order already.  This required some changes to make the string packing logic be
able to accept the sequence of strings as a pair of iterators.

Reviewed By: kodafb

Differential Revision: D14890921

fbshipit-source-id: 3fdfcbfce5d2e381d3c630b7e9183576a4586535