Factor out traversal of literal strings.
Summary:
In order to cluster string table index entries for identifiers next to each
other, I need to always eagerly gather strings from the IR (and order them).

Previously the logic that handled this was coupled with the logic that
performed the frequency analysis on strings and also the logic that packs the
string buffer.

These processes only happen when we produce optimized bytecode.  We could do
the same for gathering identifiers but run the risk of not being warned when an
identifier is missed during the initial gathering stage.  To mitigate that, I
opted to change the way strings were added to the table so that they were all
discovered before instruction selection.  This way if a string is not picked
up, when it is requested during instruction selection, it will abort due to an
assert rather than silently continue to work, but less efficiently.

This diff is the first step, separating out the traversal that needs to happen
always, from the string packing which only happens when optimizations are
enabled.

NOTE: The traversal function currently has an `includeFunctionNames` parameter.
This is temporary, and should disappear once we move to a world where all
strings are gathered before ISel.

Reviewed By: kodafb

Differential Revision: D14579041

fbshipit-source-id: 2c6156a242dd40cba8544d14132f6ed06336b4e9