Removing redundant copy/move constructor/assignment from UniquingStringLiteralTable
Summary:
This class was originally movable and not copyable because one of its data
structures pointed into another.  This is no longer true -- it has been
encapsulated in `StringSetVector` which is now movable and not copyable.

It suffices to remove the explicitly deleted and defaulted constructors and
assignment functions here -- the compiler does the right thing by inheriting
the movability and copyability of the class's fields.

Reviewed By: kodafb

Differential Revision: D14891199

fbshipit-source-id: 46da8a24c0b026d56f5e8926e43ffd8d9ff659ad