Introducing getExistingIdentifierId
Summary:
Analogous to `UniquingStringTable::getExistingStringId`, but on top of ensuring
the string in question exists in the table, also makes sure it is marked as an
identifier.

These two functions will be used to replace calls to `addString` during
instruction selection after we introduce guarantees that every string mentioned
by the bytecode will have been added beforehand.

Reviewed By: avp

Differential Revision: D14890923

fbshipit-source-id: e554097a329fa94598cd77eeb2fd05f8094fb4e6