Simplify lifecycle of UniquingStringLiteralTable
Summary:
UniquingStringTable (and now UniquingStringLiteralTable) supported generating a
storage multiple times during its lifetime.  This feature is not currently used
anywhere, and poses a problem for me:

If a table was initialised with the contents of some storage, then after the
first time it generates a storage, that information is lost.  I need this
information to always be present when generating a storage from a table,
because I want to re-order the storage's index entries as part of generation,
but only those entries corresponding to newly added strings.

This change gets rid of the ability to generate a storage multiple times so
that we never have to generate a storage after losing the information we need
to do so.

NOTE: Currently it is still possible to call `toStorage` on the same table
twice, because it accepts a reference rather than taking the table by value
(forcing it to be moved in).  The latter is the intended final state, but
currently during bytecode generation, the table is accessed after the storage
is extracted from it, so it needs to still be around.  Once I have gotten
rid of this ordering issue, I can fix this.

Reviewed By: kodafb

Differential Revision: D14892662

fbshipit-source-id: 81f758e7829524a380e1fe5cb81deb47fd45b628