Indicate when Literal Strings are being used as Identifiers.
Summary:
Augment the traversal function for `traverseLiteralStrings` with a `bool` that
indicates whether the string is being used in a way that would require it to be
treated as an "Identifier" in Hermes Bytecode parlance.

Currently we do not do anything with that information, but I will use it to
eagerly mark the strings as identifiers in a following diff.

Reviewed By: kodafb

Differential Revision: D14579040

fbshipit-source-id: 2f3dd11f789a2eb4991cdb26d6a866ae59a9e1a4