Separate FilenameTable from StringTable.
Summary:
`UniquingStringTable` is currently responsible for a number of things:

 1. Uniquing strings.
 2. Assigning IDs to strings on-demand.
 3. Loading in an existing string storage.
 4. Potentially packing strings more compactly if optimizations are enabled.
 5. Tracking which strings are identifiers (I recently moved this here from
    elsewhere).
 6. Returning the IDs of existing strings and identifiers.

`BytecodeModuleGenerator::filenameTable_` uses only features 1 and 2.  What is
more, I want to change the way string literals are accumulated so that their ID
is not available until after all strings have been captured (allowing us to
always re-order the index).  This breaks feature 2 for the filename table, it
also adds:

 7. String index re-ordering.

As a feature that `filenameTable_` supports but does not use.  All this seems
to indicate that the two tables are diverging.  All the functionality they
share is contained in either `StringSetVector` which I introduced in a previous
diff, or in `ConsecutiveStringStorage`, so I'm splitting the implementation of
`filenameTable_` into its own class and file.

Reviewed By: kodafb

Differential Revision: D14890922

fbshipit-source-id: 083d747be18e70c26259616f0762b5b9c2dcdfa6