getOrderedStringStorage takes predicate to filter functions.
Summary:
D14832014 introduces filtering into getOrderedStringStorage so that the strings
we seed the table with are limited to those in functions belonging to the
segment being compiled.  This was done by passing in the set of function
pointers we should filter down to.

I factor out the string traversal in a later diff, and it makes less sense to
pass the set of function pointers even further down, so I switched to a more
generic predicate function.

I shared the predicate with the one used by HBC to decide which function to
generate.  It differs from the one used before, which didn't return true if
the function corresponded to the entry point, and this has caused the order of
strings in the lit test to change.

Reviewed By: avp

Differential Revision: D14970874

fbshipit-source-id: e7928530dd7594a5b43ddf7b8d0827a880ad4646