Avoid conditionally private methods in MSVC
Summary:
MSVC encode public/private in the mangled name. As a result, conditionally private method leads to link failures.

The ones in this diff do not cause failure in the default build mode (as a result of selective compilation due to GC implementation switch, debug/release ifdefs, disabled tests in Windows), but should be fixed nonetheless.

Reviewed By: yinghuitan

Differential Revision: D14594605

fbshipit-source-id: 97974901d91526521cc23d318affbe0fd2d28d2f