Fix command line options for the synth trace
Summary:
The synth benchmarks need to change the defaults of some command line options.
So it can't re-use `RuntimeFlags.h` (sadly).
For now, copy the definitions.
If someone comes up with a clever re-using scheme I'll be interested.

Also, while we're doing this, change the default on `alloc-in-young` to `false`, since most synth
benchmarks want that behavior.

Reviewed By: kodafb

Differential Revision: D14979311

fbshipit-source-id: 41369d9d6fea9c5b9e670a7922ddce5bbd76760a