Fix CMake Check
Summary:
It looks like the buck target for the `cmake_check` stopped working.  My guess
is that it correlates with when the `xplat` cell disappeared.  It turns out
if it isn't looking at the correct directory, it will just fail silently with
a successful exit state.

This diff re-enables it, but also makes it accept the location of the hermes
directory as a parameter (so that it can be set differently in the buck target
and elsewhere).

It also adds a check to make sure the directory it has been given looks vaguely
like the hermes directory, by checking for `include/hermes` within it.

Finally I had to add another directory to ignore, for the synthtrace tests.

Reviewed By: dulinriley

Differential Revision: D14971446

fbshipit-source-id: e7b67d2e58d85c60d010da67ecb1e5db5cfdb721