Always collect and sort string literals before ISel.
Summary:
Simplify the handling of strings when lowering IR to HBC bytecode, while before
strings were:

 - Sometimes seeded by a traversal, sometimes not.  If not, we relied on
   strings being lazily added which were at risk of not being appropriately
   ordered.
 - Sometimes ordered to float frequently referenced strings to the top.
 - Sometimes ordered to improve compression, sometimes not.  This sorting was
   done at another time to the above, so had to be aware of it in order to
   preserve the important aspects of it.
 - Sometimes seeded by an existing string table, sometimes not.

This diff eliminates all but the last source of complexity by always collecting
strings in a traversal, and always sorting them by frequency and to improve
compression.

This was mainly achieved by splitting `UniquingStringLiteralTable` into two
parts:

 - `UniquingStringLiteralAccumulator` which collects strings without revealing
   their IDs, and then can be converted into an instance of
   `ConsecutiveStringStorage`.
 - `StringLiteralTable` which offers a way to look up the IDs for strings in
   an existing `ConsecutiveStringStorage` without the ability to add more.

`BytecodeModuleGenerator` accepts an instance of `ConsecutiveStringStorage`
and creates a `StringLiteralTable` from it.  Through the API, this forces the
pattern that all necessary strings must be accounted for before starting
bytecode generation.

`UniquingStringLiteralAccumulator` becomes the sole source of truth for string
index sorting logic.  It sorts first by frequency, and then within each
frequency class to reduce the bytecode's compressed size.  **This is why the
string table's order has changed in so many tests**:  Previously, for optimized
bytecode, strings flowed through the compiler as follows:

 - Collect string frequencies in a single traversal.
 - Sort strings by frequency and then by alphabetical order.
 - Send them through string packing.
 - Re-order index entries at the end of string-packing so that they were in
   increasing order of offset in string storage and length of string to improve
   compression performance.

In the new consolidated logic, we no longer sort in alphabetical order to break
ties from sorting by frequency, which in turn means that strings are not fed
into the packer in alphabetical order -- they go in in the order they appear in
the IR.

ridiculousfish:  Does the string packer's output have better characteristics if the
strings it gets are in something resembling alphabetical order?  If so, we may
want the packer to sort the index entries before packing.

Reviewed By: ridiculousfish

Differential Revision: D14970870

fbshipit-source-id: 9d2828eade8e5ee384a0a0d14bda7ff22465f244