Split apart string traversal function
Summary:
Split it into three parts:  One that traverses function names, another that
traverses CommonJS Module names, and a final one which traverses string
literals.

This was done because we only traverse function names if they have not been
stripped and we only traverse CommonJS module names if they have not been
statically resolved (into integer IDs).

Splitting the traversal into sections like this allows us to separate the
filtering logic from the traversal logic as much as possible.

It does also unfortunately alter the order of the traversal and so causes some
more churn in tests.

Reviewed By: kodafb

Differential Revision: D14976714

fbshipit-source-id: 463bf1dcc1decd0cbf8920ad929842829409cda4