Don't default to adding strings as not identifiers.
Summary:
kodafb mentioned in an earlier diff that it would be good to get rid of the
default value for `isIdentifier` on `UniquingStringLiteralAccumulator::addString`
so that every addition would have to make an intentional choice regarding
whether to treat the string as an identifier or not.  I agreed this was a good
idea and said I would handle it in a later diff, this is that diff.

Reviewed By: avp

Differential Revision: D14971692

fbshipit-source-id: 52697c719cb5118ec2418e94a3abf77b4513df5b