Sample a few page ids
Summary:
When bytecode I/O page tracing is enabled, log a small number of pairs of (position in access order, page id) pairs.

The data is logged using the compact base64vlq encoding, so also factor out that functionality from the SourceMap lib, so that we can use it elsewhere.

Reviewed By: ridiculousfish

Differential Revision: D15006926

fbshipit-source-id: 63bf6c5272e0b03ab19db2e61d6bc4880f13e66a