Rename per-object compaction to trimming to avoid confusion
Summary:
There was a feature of the GC that can tell objects they are going to be compacted, and they can do some extra
work to use less memory at this time.
This was called "compaction" before, but could be confusing with the general GC compaction that happens.
So now we will call it "trimming".

Reviewed By: tmikov

Differential Revision: D14953704

fbshipit-source-id: 840805375ef2a16f0ab6cf760642fd199faec158