Standardize the inGC() protocol for the sampling profiler
Summary:
The sampling profiler used to use `heapIsValid()` as a check to see if it could continue making
a sample.
`MallocGC` never invalidates the contents of the heap (it's always parseable and has data), but it *can* be moving
things around if Handle-SAN is on.

To be safe, have everything use `inGC` instead, which `MallocGC` will set at the beginning of a GC.
Add a `GCCycle` RAII class to make this easier to set reliably.

Reviewed By: davedets

Differential Revision: D14976726

fbshipit-source-id: a45d8150d22f8b7256cd00b5d88e7dd46196ed8c