simplify JSArray::putLength()
Summary:
Change `JSArray::putLength()` to use `JSObject::namedSlotRef()` and
`HermesValue::setNonPtr()`, which in turn removes the need to pass
`Runtime`. Also change the JSArray handle to a pointer.

The net result is that the new `putLength()` function is just two stores
to a fixed offset (since the length property is a known direct
property).

Reviewed By: avp

Differential Revision: D14994102

fbshipit-source-id: 2bb369110e187e1a5f4df3a4b2997a43e24548ce