Clean-up JSObject::getOwnComputedPrimitiveDescriptor
Summary:
`getOwnComputedPrimitiveDescriptor()` was always attempting to convert
its parameter to an array index, even if the object didn't have indexed
storage. It also performed some checks, which while correct, would speed
up only some rare and unusual cases while needlessly complicating the
logic.

Now the "fast path" is only for the truly fast case: we have "fast
indexed" properties and the property name already is an integer array
index.

If that fails, we proceed to look for a regular property with that name,
and finally, only if we have indexed storage, try to parse the property
name as an array index and look for it in the storage.

Reviewed By: avp

Differential Revision: D14969247

fbshipit-source-id: acd09dd2d3b7756ee38634e7864da73aa1d73b74