JSObject: remove createPropStorage()
Summary:
Previously the "indirect" property storage was passed as a constructor
parameter to JSObejct and every caller was expected to create the
indirect storage with the apropriate size in advance. Not only was that
cumbersome, but it became unnecessary when we transitioned to "direct
properties", where in most cases nothing needed to be allocated.

So, remove the concept of creating indirect storage separately and
passing it to the constructor. Instead, we switch to a model where
JSObject has methods to allocate property storage. Most of the time
these methods don't do anything and they are written in a way that when
invoked with a constant parameter for number of properties, they will
compile down to nothing.

In this way we preserve the interface's ability to create an arbitrary
number of properties, but in practice when we are creating a small
number, we do nothing.

As a side effect removed the ability to set indirect storage capacity.
That was a useless concept in this context.

Reviewed By: avp

Differential Revision: D15006511

fbshipit-source-id: a677762b5882521a694e2532f1623ac99a8ad8ff