new method JSObject::namedSlotRef()
Summary:
Introduce a new method `JSObject::namedSlotRef()` returning a reference
to a property slot, which can then be read or written to. Refactor the
existing methods to use it, with the intention to possibly deprecate
them in the future.

This allows us to use HermesValue::setNonPtr() in certain cases, which
wasn't previously possible.

This should lead to some small size wins because the new structure of
the code causes the property address to be calculated twice, but the
actual assignment or reading happens once.

Reviewed By: avp

Differential Revision: D14994101

fbshipit-source-id: 000cf4df70182c9fa4887348be023a59cae64dd0