Remove frequency class from StringPacker
Summary:
I was curious what the gains were from doing this, so I've removed it and will
see its effect on performance.  The original intention behind frequency class
was to make sure frequently accessed strings show up early in the **index** so
that they can be referred to by more compact instructions, this doesn't have to
translate to their position in the string **storage**.

I can see an argument for keeping frequently accessed strings close together
from a locality perspective, but counter by saying that hopefully in most cases
we don't have to actually read the string representation of an identifier, so
we may benefit more from having further opportunities to pack, than from the
locality benefits.

Reviewed By: ridiculousfish

Differential Revision: D14971586

fbshipit-source-id: 7f0914eee1fff3595ff2b434b4975f0a240e40c5