An IndexedEntry for every StringTableEntry
Summary:
In sorting StringTableEntries when converting aUniquingStringLiteralAccumulator
into a ConsecutiveStringStorage, I created an auxiliary structure -- a sequence
of IndexedEntries.  Because I was only sorting newly added strings, previously
I only created entries for each newly added string [0].

This diff introduces an IndexedEntry for each existing string as well, because
they are a convenient place to store the "Kind" of the associated string
(whether it is an Identifier, a Predefined Identifier, or neither) when creating
the run-length encoded representation of a string.

[0]: When I created this diff, I realised I was still reserving enough space
in the sequence to be able to store an entry for every string.  I need all that
space now, so I decided to leave it as is.

Reviewed By: kodafb

Differential Revision: D15038897

fbshipit-source-id: 858cbf2ba8779ee7a5497092f7466d57827da963