Group Predefined Identifiers together.
Summary:
Now, within each frequency class, we order all the strings (not identifiers)
together, then all the identifiers which are not predefined, and finally all
the predefined identifiers.

Note that we could have grouped **all** predefined strings (not just those used
as identifiers) together.  I chose not to because in the context of the
optimisation we are trying to perform, it could potentially cause us to add
more identifiers into the table when loading the bytecode module than before.

If non-identifier predefined strings are not accessed often during TTI, this
extra work would be wasted.  If they are, then this will avoid an inevitable
hash collision during bytecode execution and may speed things up further.  It
is worth investigating this separately.

Reviewed By: kodafb

Differential Revision: D15038896

fbshipit-source-id: b6e8f68bf361c83e1e1de8c823a78034346abe88