Add top N slowest tests to the testsuite
Summary:
Sometimes successful tests can take a long time and come close to the timeout,
but don't always exceed the timeout. It would be nice to have a visualization of tests
that came close to the limit (and as such, if they should be blacklisted to be conservative).

This diff adds test time tracking, and prints the top slowest tests at the end of the testsuite run.
It can be configured to print 0 tests if that is so desired, and it will omit the table entirely.

Interesting side-note: This can provide evidence for what a reasonable timeout value should be.

Reviewed By: avp

Differential Revision: D15090915

fbshipit-source-id: f41ba470b9f5f2fc3f15cbdb432fe32ca1ec0928