StringLiteralTable owns ConsecutiveStringStorage.
Summary:
Previously `StringLiteralTable` was initialised with a reference to a
ConsecutiveStringStorage, but did not take ownership of it.  Now it does take
ownership of it.

This change is motivated by the fact that I need to modify
`getIdentifierTranslations` so that for predefined identifiers it inserts the
Predefined Symbol ID and for other identifiers, it returns the precomputed
hash.  `StringLiteralTable` knows how to figure out whether a string is
predefined or not, using a hash lookup based on its UTF8 representation, but
the precomputed hash implementation for non-ASCII strings works on its UTF16
representation (which is how it is stored in the `ConsecutiveStringStorage`'s
character buffer).

By having the table own the storage, and (in a future diff) moving the
implementation of `getIdentifierTranslations` to the table, I can avoid
duplicating logic.

For now, `getIdentifierTranslations` is an alias on `StringLiteralTable`,
forwarding to its implementation on `ConsecutiveStringStorage`.

Reviewed By: kodafb

Differential Revision: D15038887

fbshipit-source-id: c46bdf1883f80dcb692b883a190c6844766f3cc7