No more ConsecutiveStringStorage::getStorageView
Summary:
This API existed to help unittests introspect on the contents of a storage, but
this is not necessary, `acquireStringStorage` can be used in the same way.
There was only one situation where the test needed to introspect the storage
and then continue to use it, and I had to get around that restriction for other
reasons in an earlier diff in the stack.

Reviewed By: dulinriley

Differential Revision: D15038889

fbshipit-source-id: a73ae5a5e8d3cda88b6b16b7faac0a7a672c9d26