Simplify and fix dangling StringRef in SourceMapGenerator
Summary:
This is another run at D14949738. The underlying bug is that, prior to
this fix, SourceMapGenerator uses an `std::vector<std::string>` and also a
map StringRef->index. If the vector reallocates the StringRefs can become
dangling.

Part of the complexity is the weird interaction between setSources() and
addSource().

Fix this by eliminating the SourceMapGenerator's sources vector. Have only
the map (as a StringMap), mapping from string to index. Compute the
sources vector on demand.

Also eliminate setSources(); use addSource() exclusively. This required
fixing one location where we would overwrite the existing source list.

Credit to motiz88 for finding the bug.

Reviewed By: motiz88

Differential Revision: D14966761

fbshipit-source-id: bbe830a84b12cfa2e8c93884dd59bd4077168316