Initialise PredefinedString Set in a loop
Summary:
The macro expansion for initialising the predefined string set previously
emitted an insertion for every predefined string.  This led to a rather large
(41KB) function, which caused performance issues.

This diff creates static arrays containing the string contents, lengths and
Predefined enum values for each predefined string and then uses them
to initialise the mapping in a loop.

Reviewed By: tmikov

Differential Revision: D15119931

fbshipit-source-id: 94941ff0c9e6cf11e5742fb0ea8c8704458d1d86