Flag dangling HermesPointerValues when JSI-side invalidates them
Summary:
Before this diff we had an assert on destruction of a HermesRuntime to catch lingering external references into the managed JS heap via HermesPointerValue instances. However, this does not give an useful information on what is actually leaking. This diff moves the to destruction of the dangling reference in JSI so we can see what held the reference.

This implementation should introduces no new significant memory overhead, by reusing the upper bits of the HermesPointerValue reference count to flag dangling references when HermesRuntime is destroyed. This implementation assumes we'll never have more than 2^29 references to any one HermesPointerValue. See the in-line comments for details of why we need 2 bits.

In the case of a dangling reference we deliberately leak some heap memory. This should be fine as the whole process should (hopefully) be killed soon by an assertion  showing us where the dangling pointer is.

Reviewed By: mhorowitz

Differential Revision: D15178670

fbshipit-source-id: c0777383273d0009b660f53d520fdad5ebf594f2